Bug 305397 -  Failing test in I-1.2.0-20100304170940 build

+++ bundles/org.eclipse.wst.jsdt.debug.core/src/org/eclipse/wst/jsdt/debug/internal/core/JavaScriptPreferencesManager.java
@@ -62,7 +62,9 @@ public class JavaScriptPreferencesManager implements IPreferenceChangeListener {
 	public void stop() {
 		new InstanceScope().getNode(JavaScriptDebugPlugin.PLUGIN_ID).removePreferenceChangeListener(this);
 		try {
-			allLoadsBreakpoint.delete();
+			if(allLoadsBreakpoint != null) {
+				allLoadsBreakpoint.delete();
+			}
 		} catch (CoreException e) {
 			JavaScriptDebugPlugin.log(e);
 		}
